Chopsum
Super Weird Tech Company recently lost a big amount of revenue due to several incidents of business documents and messages getting tampered. One of the most effective ways to detect these types of tampering is the checksum. As an employee of Super Weird Tech Company, you have the responsibility of developing a checksum algorithm that is reasonably fast. At first, you thought of counting all of the unique characters in the string. But, this technique has a scary flaw. Someone can still manipulate the non-repeating characters and the checksum will not be changed. Hence, the modification will be unnoticed. To fix this problem, you chopped the message in each and every possible way possible and accumulated the number of non-repeating letters. Now, this accumulated result is the desired checksum. In this way, even slight tampering in any portion of the message can be detected.

Now, you need to develop a solution to calculate the checksum in a reasonable time. As the result can be huge, it should be checksum modulo 109+7 to prevent overflowing issues.

Input:
First line will contain T, number of testcases. Then the testcases follow.
Each testcase contains of a single line of input, the message string.
Output:
For each test case, output in a single line answer given by Checksum of message modulo 1000000007.

Constraints
0≤Message.length≤104
Message only has ASCII-128 printable characters
Sample Input:
1
1RR
Sample Output:
6
EXPLANATION:
All possible substrings are: "1","R","R","1R","RR" and "1RR" So, the checksum is : 1 + 1 + 1 + 2 + 0 + 1 = 6

Unable to View it correctly?
LInk to Problem Statement: https://www.codechef.com/problems/HB820CHS